Fixed I2C stuck when attempting transmission

+++ Src/HAL/I2CInterface.hh
@@ -11,7 +11,7 @@
 
 class I2CInterface {
     virtual void readBytes(uint8_t startingAddress, uint8_t consecutiveBytes, uint8_t *outputArray) = 0;
-    virtual void writeByte(uint8_t targetAddress, uint8_t data) = 0;
+    virtual void writeByte(uint16_t targetAddress, uint8_t data) = 0;
     virtual uint8_t readByte(uint8_t address) = 0;
     virtual void writeBytes(uint8_t startingAddress, uint16_t numberOfBytes, uint8_t *dataToWrite) = 0;
 };
